+++ app/models/note/skos/change_note.rb
@@ -21,6 +21,17 @@ class Note::SKOS::ChangeNote < Note::SKOS::Base
   def self.edit_partial_name(obj)
     "partials/note/skos/edit_change_note"
   end
+  
+  def self.search_result_partial_name
+    'partials/note/skos/change_note/search_result'
+  end
+  
+  def self.single_query(params = {})
+    query_str = build_query_string(params)
+
+    Note::SKOS::ChangeNote.includes(:annotations).
+      merge(Note::Annotated::Base.where(Note::Annotated::Base.arel_table[:value].matches(query_str)))
+  end
 
   def build_rdf(document, subject)
     subject.send(self.rdf_namespace).build_predicate(self.rdf_predicate) do |blank_node|